requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiogram-tests==1.0.3
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
deepl==1.15.0
exceptiongroup==1.1.3
flake8==6.1.0
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
iniconfig==2.0.0
linecache2==1.0.0
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
pycodestyle==2.11.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyflakes==3.1.0
pytest==7.4.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.21
tomli==2.0.1
traceback2==1.4.0
typing_extensions==4.7.1
unittest2==1.1.0
urllib3==2.0.6
yarl==1.9.2
